+++ daimaduan/config.cfg.sample
@@ -1,4 +1,4 @@
-[mongodb]
+[mongoengine]
 host = 127.0.0.1
 database = daimaduan
