+++ daimaduan/tasks/seed_data.py
@@ -1,7 +1,7 @@
 from fabric.decorators import task
 from fabric.operations import local
-from daimaduan.models import Syntax
 
+from daimaduan.models import Syntax
 
 syntax_list = [
     "CSS",
