--- daimaduan/utils.py
@@ -1,182 +1,0 @@
-# coding: utf-8
-
-"""
-    daimaduan.utils
-    ---------------
-
-    A set of utilities.
-"""
-import logging
-import math
-from functools import wraps
-from json import dumps
-
-from bottle import jinja2_template
-from bottle import request
-from bottle import response
-from bottle_utils.csrf import generate_csrf_token
-from itsdangerous import URLSafeTimedSerializer
-from mailthon import email
-from mailthon.middleware import Auth
-from mailthon.middleware import TLS
-from mailthon.postman import Postman
-
-from daimaduan.models import User
-from daimaduan.models import UserOauth
-
-
-# Setup logger
-logging.basicConfig(format='%(levelname)s %(asctime)s %(message)s', level=logging.INFO)
-logger = logging.getLogger('daimaduan')
-
-
-def oauth_config(config, provider):
-    return {
-        'name': config['oauth.%s.name' % provider],
-        'client_id': config['oauth.%s.client_id' % provider],
-        'client_secret': config['oauth.%s.client_secret' % provider],
-        'authorize_url': config['oauth.%s.authorize_url' % provider],
-        'access_token_url': config['oauth.%s.access_token_url' % provider],
-        'base_url': config['oauth.%s.base_url' % provider]
-    }
-
-
-def get_session(request):
-    """Get session instance from request"""
-
-    return request.environ.get('beaker.session')
-
-
-def user_bind_oauth(user, session):
-    """Bind oauth info in session to given user, then clear the session"""
-
-    oauth = UserOauth(user=user,
-                      provider=session['oauth_provider'],
-                      openid=session['oauth_openid'],
-                      token=session['oauth_token'])
-    oauth.save()
-
-    # Clear oauth info in session
-    del session['oauth_provider']
-    del session['oauth_openid']
-    del session['oauth_name']
-    del session['oauth_token']
-
-
-def jsontify(func):
-    @wraps(func)
-    def jsontify_function(*args, **kwargs):
-        result = func(*args, **kwargs)
-        response.content_type = 'application/json'
-        return dumps(result)
-    return jsontify_function
-
-
-def generate_confirmation_token(config, email):
-    """generate confirmation token using user's email via itsdangerous"""
-    serializer = URLSafeTimedSerializer(config['site.secret_key'])
-    return serializer.dumps(email, salt=config['site.token_salt'])
-
-
-def validate_token(config, token, expire_time=3600):
-    """from token and expire_time to confirm user's email"""
-    serializer = URLSafeTimedSerializer(config['site.secret_key'])
-    try:
-        confirmed_email = serializer.loads(token, max_age=expire_time, salt=config['site.token_salt'])
-    except Exception:
-        return False
-    return confirmed_email
-
-
-SENDER = 'noreply.daimaduan@gmail.com'
-CONFIRMATION_SUBJECT = u'请激活您在代码段注册的邮箱地址'
-CONFIRMATION_CONTENT = u"<p>你好 %s,</p> \
-    <p>非常感谢注册代码段. 请点击<a href=\"https://%s/confirm/%s\">链接</a>来激活您的邮箱地址.</p> \
-            <p>或者您可以拷贝下面这个地址到您的浏览器中访问来激活邮箱地址 https://%s/confirm/%s </p> \
-            <br> \
-            <p>此致！</p> \
-            <p>代码段团队</p>"
-RESET_PASSWORD_SUBJECT = u'您请求重置密码'
-RESET_PASSWORD_CONTENT = u"<p>你好 %s,</p> \
-    <p>您已请求重置代码段密码. 请点击<a href=\"https://%s/reset_password/%s\">链接</a>来重置您的密码.</p> \
-            <p>或者您可以拷贝下面这个地址到您的浏览器中访问来重置您的密码 https://%s/reset_password/%s </p> \
-            <br> \
-            <p>此致！</p> \
-            <p>代码段团队</p>"
-
-
-def send_confirm_email(config, user_email):
-    token = generate_confirmation_token(config, user_email)
-    user = User.objects(email=user_email).first()
-    content = CONFIRMATION_CONTENT % (user.username, config['site.domain'], token, config['site.domain'], token)
-    send_email(config, user_email, CONFIRMATION_SUBJECT, content)
-
-
-def send_reset_password_email(config, user_email):
-    token = generate_confirmation_token(config, user_email)
-    user = User.objects(email=user_email).first()
-    content = RESET_PASSWORD_CONTENT % (user.username, config['site.domain'], token, config['site.domain'], token)
-    send_email(config, user_email, RESET_PASSWORD_SUBJECT, content)
-
-
-def send_email(config, user_email, subject, content):
-    envelope = email(
-        sender=SENDER,
-        receivers=[user_email],
-        subject=subject,
-        content=content
-    )
-
-    postman = Postman(host=config['mail.host'],
-                      port=int(config['mail.port']),
-                      middlewares=[TLS(force=True),
-                                   Auth(username=config['mail.username'], password=config['mail.password'])])
-    try:
-        response = postman.send(envelope)
-        if response.ok:
-            logger.info("Successfully send email to %s" % user_email)
-        else:
-            logger.error("Failed to send email to %s" % user_email)
-    except Exception, e:
-        logger.error("Exception occured when send email to %s" % e)
-
-
-def user_active_required(func):
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        if request.user.is_email_confirmed:
-            return func(*args, **kwargs)
-        generate_csrf_token()
-        return jinja2_template('email/active.html', email=request.user.email, title=u"邮箱需要激活", reactive=True, token=request.csrf_token)
-    return wrapper
-
-
-def paginate(collection, current_page, per_page=20):
-    total_page = int(math.ceil((float(collection.count()) / per_page)))
-
-    start_index = (current_page - 1) * per_page
-    end_index = current_page * per_page
-    page_collection = collection[start_index:end_index]
-
-    summary = {'total_page': total_page,
-               'current_page': current_page,
-               'previous_page': current_page - 1,
-               'next_page': current_page + 1,
-               'is_first_page': current_page == 1,
-               'is_last_page': current_page == total_page}
-
-    return page_collection, summary
-
-
-def get_page():
-    """ Get current page
-
-    Get current page from query string `page=x`,
-    if page not given returns `1` instead.
-    """
-
-    try:
-        page = request.params.get('page', 1)
-        return int(page)
-    except ValueError:
-        return 1
