+++ daimaduan/utils/commons.py
@@ -1,0 +1,4 @@
+def get_session(request):
+    """Get session instance from request"""
+
+    return request.environ.get('beaker.session')
